Vet clinic database: add database schema diagram
Make a database schema diagram for the vet clinic database you have been creating throughout the projects.
Remember to include the diagram in your repo.
You need to use pair-programming for this project. Therefore you need to implement the above features in all of the repos in your pair-programming group.